+++ src/Monitor/ActivityLogAlert.Autorest/Properties/AssemblyInfo.cs
@@ -24,5 +24,5 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 [assembly: CLSCompliant(false)]
 [assembly: Guid("68c9e5fe-3c10-46ed-b371-9a8c93ad49c7")]
-[assembly: AssemblyVersion("4.6.0")]
-[assembly: AssemblyFileVersion("4.6.0")]
+[assembly: AssemblyVersion("5.0.0")]
+[assembly: AssemblyFileVersion("5.0.0")]